src/main/java/com/example/demotdd/ProductController.java
package com.example.demotdd;

import com.example.demotdd.entity.Product;
import lombok.NoArgsConstructor;
import lombok.RequiredArgsConstructor;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("products")
@RequiredArgsConstructor
public class ProductController {

    private final ProductService productService;

    @GetMapping("{sku}")
    public Product getBySku(@RequestParam String sku) {
        return productService.getBySku(sku);
    }

}
